Instead of transmitting strings back and forth, use booleans
Here you've pushed a 'no' or an enzyme name back and forth--the enzyme name doesn't
even need to be passed to the function. So, since it really boils down to a simple if/else, true/false
at the end of the function, we just return that. It makes simplifies  section as well